+++ sql-planner/src/executor/engine/helpers.rs
@@ -1166,9 +1166,10 @@ pub fn materialize_values(
         runtime.dispatch(exec_plan, values_id, &Buckets::Any, &mut port)?;
 
         let mut vtable = VirtualTable::with_columns(columns);
-        let mut ys = Scheduler::default();
+        let scheduler_opts = runtime.get_scheduler_options();
+        let mut ys = Scheduler::new(&scheduler_opts);
         for mp in port.iter().skip(1) {
-            ys.maybe_yield(&runtime.get_scheduler_options())
+            ys.maybe_yield(&scheduler_opts)
                 .map_err(|e| SbroadError::Other(e.to_smolstr()))?;
             vtable.write_all(mp).map_err(|e| {
                 SbroadError::FailedTo(
@@ -1293,9 +1294,10 @@ pub fn materialize_motion(
             })?;
         }
     } else {
-        let mut ys = Scheduler::default();
+        let scheduler_opts = runtime.get_scheduler_options();
+        let mut ys = Scheduler::new(&scheduler_opts);
         for mp in port.iter().skip(1) {
-            ys.maybe_yield(&runtime.get_scheduler_options())
+            ys.maybe_yield(&scheduler_opts)
                 .map_err(|e| SbroadError::Other(e.to_smolstr()))?;
             vtable.write_all(mp).map_err(|e| {
                 SbroadError::FailedTo(
@@ -1431,7 +1433,7 @@ pub fn old_populate_table(
                 )),
             )
         })?;
-        let mut ys = Scheduler::default();
+        let mut ys = Scheduler::new(options);
         for tuple in data.iter() {
             ys.maybe_yield(options)
                 .map_err(|e| SbroadError::Other(e.to_smolstr()))?;