+++ test/int/test_sql.py
@@ -667,6 +667,7 @@ def test_read_from_system_tables(cluster: Cluster):
         "sql_motion_row_max",
         "sql_preemption",
         "sql_preemption_interval_us",
+        "sql_preemption_opcode_max",
         "sql_storage_cache_count_max",
         "sql_storage_cache_size_max",
         "sql_vdbe_opcode_max",